+++ examples/lxmf_echobot.js
@@ -1,11 +1,5 @@
-import {
-  fromHex,
-  Identity,
-  LXMessage,
-  LXMRouter,
-  Reticulum,
-  toHex,
-} from "@reticulum/core";
+import { fromHex, Identity, Reticulum, toHex } from "@reticulum/core";
+import { LXMessage, LXMRouter } from "@reticulum/core/src/lxmf/index.js";
 import {
   FileStorageAdapter,
   LocalClientInterface,